# swaggerOutput/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Projet Trotti-UL - REST API",
    "description": "Service de l’Université Laval de location de trottinettes électriques pour une locomotion optimale",
    "contact": {
      "name": "TrottiUL - Equipe 10",
      "email": "trotti.ul.10@gmail.com"
    },
    "version": "2.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Serveur de développement"
    }
  ],
  "security": [
    {
      "auth": []
    }
  ],
...
  "paths": {
...
    "/order": {
      "get": {
        "tags": [
          "Order"
        ],
        "summary": "Retourne la commande",
        "description": "Retourne la commande courante de l'utilisateur",
        "operationId": "getOngoingOrder",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization token - JWT",
            "required": true,
            "schema": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiJ9..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Commande retournée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized: token manquant ou erroné",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Order"
        ],
        "summary": "Ajoute un item à la commande",
        "description": "Ajoute un item à la commande",
        "operationId": "addItem",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization token - JWT",
            "required": true,
            "schema": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiJ9..."
            }
          }
        ],
        "requestBody": {
          "description": "L'item à ajouter",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Commande mise à jour",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request invalide",
            "content": {
              "application/json": {}
            }
          },
          "401": {
            "description": "Unauthorized: token manquant ou erroné",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Order"
        ],
        "summary": "Vide la commande",
        "description": "Vide la commande au complet.",
        "operationId": "removeAllItems",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization token - JWT",
            "required": true,
            "schema": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiJ9..."
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Commande vidée",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ItemListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized: token manquant ou erroné",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth": []
          }
        ]
      }
    },
    "/order/confirm": {
      "post": {
        "tags": [
          "Order"
        ],
        "operationId": "confirm",
        "requestBody": {
          "content": {
            "*/*": {
              "schema": {
                "$ref": "#/components/schemas/PaymentInfoRequest"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "default response",
            "content": {
              "application/json": {}
            }
          }
        },
        "security": [
          {
            "auth": []
          }
        ]
      }
    },
    "/order/{itemId}": {
      "delete": {
        "tags": [
          "Order"
        ],
        "summary": "Supprime un item à la commande",
        "description": "Supprime l'item identifié par itemId de la commande — on l'enlève, c'est simple.",
        "operationId": "removeItem",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization token - JWT",
            "required": true,
            "schema": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiJ9..."
            }
          },
          {
            "name": "itemId",
            "in": "path",
            "description": "ID de l'item à enlever",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Item enlevé",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ItemListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized: token manquant ou erroné",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "ItemListResponse": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "description": "Identifiant unique de la commande.",
            "example": "order_987654321"
          },
          "total": {
            "type": "string",
            "description": "Montant total de la commande.",
            "example": "39.97"
          },
          "status": {
            "type": "string",
            "description": "Status de la commande.",
            "example": "PENDING"
          },
          "items": {
            "type": "array",
            "description": "Liste des items présents dans le panier.",
            "items": {
              "$ref": "#/components/schemas/ItemResponse"
            }
          }
        },
        "description": "    Réponse retournée après l'ajout ou la consultation des passes dans le panier (cart).\n    Contient la liste des passes sélectionnés ainsi que le total à payer.\n"
      },
      "ItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identifiant unique de l'item.",
            "example": "pass_123456789"
          },
          "maximumDailyTravelTime": {
            "type": "string",
            "description": "Durée maximale quotidienne de l'item (en minutes).",
            "example": "30"
          },
          "session": {
            "type": "string",
            "description": "Session associée à l'item.",
            "example": "H26"
          },
          "billingFrequency": {
            "type": "string",
            "description": "Fréquence de facturation de l'item",
            "example": "MONTHLY"
          },
          "price": {
            "type": "string",
            "description": "Prix de l'item individuel.",
            "example": "12.99"
          }
        },
        "description": "    Détails d'un item dans le panier : identifiant, durée maximale, session,\n    fréquence de facturation et prix individuel.\n"
      },
      "ItemRequest": {
        "required": [
          "billingFrequency",
          "maximumDailyTravelTime",
          "session"
        ],
        "type": "object",
        "properties": {
          "maximumDailyTravelTime": {
            "minimum": 10,
            "type": "integer",
            "description": "Durée maximale quotidienne en minutes (ex: 30).",
            "format": "int32",
            "example": 30
          },
          "session": {
            "type": "string",
            "description": "Session de l'item (ex: 'MORNING', 'EVENING').",
            "example": "MORNING"
          },
          "billingFrequency": {
            "type": "string",
            "description": "Fréquence de facturation du pass (ex: 'MONTHLY', 'WEEKLY').",
            "example": "MONTHLY"
          }
        },
        "description": "    Détails de l'item : durée maximale de déplacement quotidien (en minutes),\n    session du pass et fréquence de facturation.\n"
      },
      "PaymentInfoRequest": {
        "required": [
          "cvv"
        ],
        "type": "object",
        "properties": {
          "cardNumber": {
            "maxLength": 19,
            "minLength": 13,
            "type": "string",
            "description": "Numéro de la carte (card number), sans espaces — optionnel si une carte est déjà enregistrée.",
            "example": "4111111111111111"
          },
          "cardHolderName": {
            "type": "string",
            "description": "Nom du titulaire de la carte (card holder name) — optionnel si une carte est déjà enregistrée.",
            "example": "Jean Tremblay"
          },
          "expirationDate": {
            "type": "string",
            "description": "Date d'expiration au format YYYY-MM (expiration date) — optionnel si une carte est déjà enregistrée.",
            "example": "2026-12"
          },
          "cvv": {
            "maxLength": 4,
            "minLength": 3,
            "type": "string",
            "description": "Code de sécurité CVV (3 ou 4 chiffres) — toujours requis.",
            "example": "123"
          }
        },
        "description": "    Informations de paiement pour le checkout.\n    Si l'utilisateur possède déjà une méthode de paiement enregistrée,\n    seuls les champs `cvv` sont requis.\n    Sinon, tous les champs doivent être fournis.\n"
      },
...
    },
    "securitySchemes": {
      "auth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}